requirements.txt
flask
selenium
webdriver-manager
beautifulsoup4
mysql.connector